+++ src/plugin-api/main/uk/co/reecedunn/intellij/plugin/processor/log/LogViewProvider.kt
@@ -21,4 +21,6 @@ interface LogViewProvider : QueryProcessor {
     fun logs(): List<String>
 
     fun log(name: String): List<String>
+
+    fun defaultLogFile(logs: List<String>): String?
 }